literal/requirements.txt
Django==3.2.2
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.8.6
pydantic==1.8.1
